chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It includes updating `mongoose` to version 8.4.3, `next` to version 14.2.4, and `nodemailer` to version 6.9.14. Additionally, it updates the dev dependencies, including `@types/nodemailer` to version 6.4.15, `@types/react` and `@types/react-dom` to version 18, `eslint` to version 8, `eslint-config-next` to version 14.2.4, and `typescript` to version 5.