chore: Update npm dependencies
This commit updates the npm dependencies in the `package.json` file. It includes updating `bcryptjs` to version 2.4.3, `jsonwebtoken` to version 9.0.2, `mongoose` to version 8.4.3, `next` to version 14.2.4, and `nodemailer` to version 6.9.14. Additionally, it updates the dev dependencies, including `@types/bcryptjs` to version 2.4.6, `@types/jsonwebtoken` to version 9.0.6, `@types/node` to version 20, `@types/nodemailer` to version 6.4.15, and `@types/react` to version 18.